feat: add check if knuu was initialized

+++ pkg/k8s/k8s_pod.go
@@ -1,22 +1,22 @@
 package k8s
 
 import (
-	"bytes"
-	"context"
-	"fmt"
-	"io"
-	"k8s.io/client-go/tools/portforward"
-	"k8s.io/client-go/transport/spdy"
-	"net/http"
-	"k8s.io/apimachinery/pkg/api/resource"
-	"strings"
-	"time"
+    "bytes"
+    "context"
+    "fmt"
+    "io"
+    "k8s.io/apimachinery/pkg/api/resource"
+    "k8s.io/client-go/tools/portforward"
+    "k8s.io/client-go/transport/spdy"
+    "net/http"
+    "strings"
+    "time"
 
-	"github.com/sirupsen/logrus"
-	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/tools/remotecommand"
+    "github.com/sirupsen/logrus"
+    v1 "k8s.io/api/core/v1"
+    metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+    "k8s.io/client-go/kubernetes/scheme"
+    "k8s.io/client-go/tools/remotecommand"
 )
 
 // getPod retrieves a pod from the given namespace and logs any errors.
@@ -25,6 +25,9 @@ func getPod(namespace, name string) (*v1.Pod, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), 20*time.Second)
 	defer cancel()
 
+	if !IsInitialized() {
+		return nil, fmt.Errorf("knuu is not initialized")
+	}
 	pod, err := Clientset().CoreV1().Pods(namespace).Get(ctx, name, metav1.GetOptions{})
 	if err != nil {
 		return nil, fmt.Errorf("failed to get pod %s: %w", name, err)
@@ -45,6 +48,9 @@ func DeployPod(podConfig PodConfig, init bool) (*v1.Pod, error) {
 	defer cancel()
 
 	// Try to create the pod
+	if !IsInitialized() {
+		return nil, fmt.Errorf("knuu is not initialized")
+	}
 	createdPod, err := Clientset().CoreV1().Pods(podConfig.Namespace).Create(ctx, pod, metav1.CreateOptions{})
 	if err != nil {
 		return nil, fmt.Errorf("failed to create pod: %v", err)
@@ -126,6 +132,9 @@ func RunCommandInPod(namespace, podName, containerName string, cmd []string) (st
 	}
 
 	// Construct the request for executing the command in the specified container
+	if !IsInitialized() {
+		return "", fmt.Errorf("knuu is not initialized")
+	}
 	req := Clientset().CoreV1().RESTClient().Post().
 		Resource("pods").
 		Name(podName).
@@ -185,6 +194,9 @@ func DeletePod(namespace, name string) error {
 	defer cancel()
 
 	// Delete the pod using the Kubernetes client API
+	if !IsInitialized() {
+		return fmt.Errorf("knuu is not initialized")
+	}
 	if err := Clientset().CoreV1().Pods(namespace).Delete(ctx, name, metav1.DeleteOptions{}); err != nil {
 		return fmt.Errorf("failed to delete pod %s: %v", name, err)
 	}
@@ -408,6 +420,9 @@ func PortForwardPod(namespace string, podName string, localPort int, remotePort 
 	}
 
 	// Setup the port forwarding
+	if !IsInitialized() {
+		return fmt.Errorf("knuu is not initialized")
+	}
 	url := Clientset().CoreV1().RESTClient().Post().
 		Resource("pods").
 		Namespace(namespace).
